requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.3.0
Pillow==9.2.0
pygments==2.11.2
django-filter==21.1